Spit tests by root folder #69
