+++ tests/integration/default-root.js
@@ -4,7 +4,6 @@ var Mugshot = require('../../lib/mugshot.js');
 var WdioAdapter = require('mugshot-webdriverio');
 var path = require('path');
 var fs = require('fs');
-var tmp = require('tmp');
 
 const URL = 'file://' + path.join(__dirname, 'test.html');
 
@@ -17,7 +16,6 @@ const BROWSER_OPTIONS = {
 const name = 'great',
       ext = '.png',
       dir = 'visual-tests',
-      extendedDir = 'visual-tests-extended/components/rectangle',
       paths = [path.join(dir, name + ext),
                path.join(dir, name + '.new' + ext),
                path.join(dir, name + '.diff' + ext)];
@@ -181,50 +179,6 @@ describe('Mugshot integration with disabled acceptFirstBaseline', function() {
 
   after(function() {
     cleanUp();
-    return wdioInstance.end();
-  });
-});
-
-
-describe('Mugshot integration with recursive rootDirectory', function() {
-  this.timeout(0);
-  var wdioInstance, browser;
-  before(function() {
-
-    return wdioInstance = wdio.remote(BROWSER_OPTIONS).init().url(URL)
-        .then(function() {
-          browser = new WdioAdapter(this);
-        });
-  });
-
-  it('should create the desired rootDirectory', function(done) {
-
-    tmp.dir({unsafeCleanup: true},
-        function _tempDirCreated(error, tempPath) {
-          if (error) {
-            throw error;
-          }
-
-          var mugshotOptions = {
-                rootDirectory: path.join(tempPath, extendedDir)
-              },
-              expected = {
-                isEqual: true,
-                baseline: path.join(tempPath, extendedDir,
-                    noDifferencesSelector.name + ext)
-              },
-              mugshot = new Mugshot(browser, mugshotOptions);
-
-          mugshot.test(noDifferencesSelector, function(error, result) {
-            expect(error).to.be.null;
-            expect(result).to.be.deep.equal(expected);
-
-            done();
-          });
-        });
-  });
-
-  after(function() {
     return wdioInstance.end();
   });
 });
